% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animateFX.R
\name{animateFX}
\alias{animateFX}
\title{Animate PITCHf/x}
\usage{
animateFX(data, pitch_interval = 0.01, layers = facet_grid(~stand),
  color = "pitch_type", point.alpha = 1/3, x_lim = c(-3.5, 3.5),
  y_lim = c(0, 7), movie.name = "animation.gif", ...)
}
\arguments{
\item{data}{data frame with appropriately named PITCHf/x variables}

\item{layers}{list of ggplot2 layer modifications.}

\item{color}{character. plot point color.}

\item{point.alpha}{ggplot2 alpha parameter}

\item{x_lim}{vector(2 element). x axis range.}

\item{y_lim}{vector(2 element). y axis range.}

\item{movie.name}{gif image name}

\item{...}{using in saveGIF}

\item{interval}{time (in seconds) between plotting the pitch locations.}
}
\value{
Returns a series of objects of the class used by package ggplot2 to represent plots.
}
\description{
Pitch trajectories animated on a two-dimensional plot.
}
\details{
\code{animateFX} plots a series of "snapshots" that represent pitch trajectories
from the point of release until all of them reach home plate.
The graphic takes on the viewpoint of the umpire; that is, the pitches are getting closer
to the viewer with time. This is relected with the increase in size of the "balls" as the
animation progresses.
}
\examples{
\dontrun{
  animateFX(pitches, interval = 1.0, ani.height = 600, ani.width = 1200)
}

}
