% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snapshots.R
\name{get_snapshots}
\alias{get_snapshots}
\title{Produce time sequenced pitch locations from PITCHf/x parameters}
\usage{
get_snapshots(data, interval = 0.01)
}
\arguments{
\item{data}{The nine PITCHf/x parameters used to determine the location of a pitch at a given time.}

\item{interval}{the amount of time between 'snapshots'}
}
\value{
Return a three dimensional array. The third dimension corresponds to different 'snapshots' of locations.
}
\description{
Produce time sequenced pitch locations from PITCHf/x parameters
}
\references{
http://baseball.physics.illinois.edu/KaganPitchfx.pdf
}
