% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strikezones.R
\name{get_strikezones}
\alias{get_strikezones}
\title{Calculate strikezone boundaries}
\usage{
get_strikezones(data, mean_strikezone = FALSE)
}
\arguments{
\item{data}{PITCHf/x orginally entered into \code{animateFX}}
}
\value{
Returns a list of boundaries for both right handed batters and left handed batters
}
\description{
Strikezone boundaries calculated according to Mike Fast's specifications
}
\references{
\url{http://www.baseballprospectus.com/article.php?articleid=14572}
}
