% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{make data which is easy to use}
\usage{
make_data(db_sqlite, player)
}
\arguments{
\item{db_sqlite}{sqlite file.}

\item{player}{data.frame which has id and name column.}
}
\value{
data.frame
}
\description{
make data which is easy to use
}
\examples{
make_data("mlb2018.sqlite")

}
