% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_game_info.R
\name{scrape_game_info}
\alias{scrape_game_info}
\title{Scrape MLB Players Data}
\usage{
scrape_game_info(gids)
}
\arguments{
\item{gids}{Gameday URLs vector.}
}
\value{
list
}
\description{
Function for obtaining MLB Players Data.
\href{http://gd2.mlb.com/components/game/mlb/year_2019/month_04/day_04/gid_2019_04_04_wasmlb_nynmlb_1/players.xml}{players.xml}
}
\examples{
data(game_ids, package = "pitchRx2")
gid <- str_subset(game_ids, "^gid_2019_04_05_")
scrape_game_info(gid)
}
