% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animateFX.R
\name{animateFX}
\alias{animateFX}
\title{Animate PITCHf/x}
\usage{
animateFX(data, point.alpha = 1/3, layers, strike = FALSE,
  color = "pitch_type_name", x_lim = c(-3.5, 3.5), y_lim = c(0, 7),
  flag = FALSE, interval = 0.01, ...)
}
\arguments{
\item{data}{data frame with appropriately named PITCHf/x variables}

\item{point.alpha}{ggplot2 alpha parameter}

\item{color}{variable used to control coloring scheme.}

\item{flag}{indicate whether or not batter has decided to swing.}

\item{interval}{time (in seconds) between plotting the pitch locations.}

\item{...}{extra options passed onto geom commands}

\item{avg.by}{variable used as an index for averaging over PITCHf/x parameters}

\item{limitz}{limits for horizontal and vertical axes.}

\item{layer}{list of ggplot2 layer modifications.}

\item{parent}{is the function being called from a higher-level function? (experimental)}
}
\value{
Returns a series of objects of the class used by package ggplot2 to represent plots.
}
\description{
Pitch trajectories animated on a two-dimensional plot.
}
\details{
\code{animateFX} plots a series of "snapshots" that represent pitch trajectories
from the point of release until all of them reach home plate.
The graphic takes on the viewpoint of the umpire; that is, the pitches are getting closer
to the viewer with time. This is relected with the increase in size of the "balls" as the
animation progresses.
}
\examples{
data(pitches)
#generate animation and prompt default web browser to view the sequence of plots
\dontrun{
animation::saveHTML({ animateFX(pitches, layer = facet_grid(pitcher_name~stand)) })
animation::saveHTML({ animateFX(pitches, avg.by="pitch_types",
                         layer = facet_grid(pitcher_name~stand))
                   })
}

}
