% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_players.R
\name{scrape_players}
\alias{scrape_players}
\title{Scrape MLB Players Data}
\usage{
scrape_players(gids)
}
\arguments{
\item{gids}{Gameday URLs vector.}
}
\value{
list
}
\description{
Function for obtaining MLB Players Data.
\href{http://gd2.mlb.com/components/game/mlb/year_2011/month_04/day_04/gid_2011_04_04_minmlb_nyamlb_1/players.xml}{players.xml}
}
\examples{
data(game_ids, package = "pitchRx2")
gid <- str_subset(game_ids, "^gid_2017_04_05_")
scrape_players(gid)

}
